Solution for Margins and Padding Ex 02
Give 8px margin between the edge of the card and its content
  (the blue sections).
Give an 8px gap between each of the blue sections inside the card.
Give the title of the card a 16px font.
Give 8px padding for the title text
Give the content section 16px margin top and bottom and 8px on either
  side.
Make everything inside the `.button-container` section centered, and
  give it 8px padding.
Center the Big Button
Give the Big Button 24px padding on the sides and 8px on top and bottom.