Solution to CSS Exercises 06-flex-layout
Create header at the top of the page, the footer at the bottom, and make
  sure they stay in place if you resize your screen.
Give the header and footer padding.
Push the links in the header and footer to either side.
Give space between the links in the header and footer.
Give the footer a light gray background (`#eeeeee`).
Center the logo, input and buttons
Give the buttons an appropriate amount of padding.
Make sure there is space between the logo, input and buttons.